Expose API to user for reopening log files
`init_file` and `init_config` both return a `Handle` on success. The
handle can be used to call reopen which runs the `post_rotate` function
on each appender.

The file reopen test was moved from the file module itself to an
integration test that makes use of the log macros. Aside from more code
coverage, the test should be less sensitive to changes in implementation
details.